# backend/follower-service/src/main/java/com/socialblog/follower/api/FollowController.java
package com.socialblog.follower.api;import com.socialblog.follower.application.FollowService;import com.socialblog.follower.application.FollowService.*;import org.springframework.data.domain.*;import org.springframework.security.core.annotation.AuthenticationPrincipal;import org.springframework.security.oauth2.jwt.Jwt;import org.springframework.web.bind.annotation.*;import java.util.UUID;@RestController @RequestMapping("/api/v1/follows")public class FollowController{private final FollowService service;public FollowController(FollowService s){service=s;}@PutMapping("/{target}")Relationship follow(@PathVariable UUID target,@AuthenticationPrincipal Jwt j){return service.follow(user(j),target);}@DeleteMapping("/{target}")Relationship unfollow(@PathVariable UUID target,@AuthenticationPrincipal Jwt j){return service.unfollow(user(j),target);}@GetMapping("/status/{target}")Relationship status(@PathVariable UUID target,@AuthenticationPrincipal Jwt j){return service.status(user(j),target);}@GetMapping("/users/{id}/followers")Page<UserLink>followers(@PathVariable UUID id,@RequestParam(defaultValue="0")int page,@RequestParam(defaultValue="20")int size){return service.followers(id,PageRequests.of(page,size,100));}@GetMapping("/users/{id}/following")Page<UserLink>following(@PathVariable UUID id,@RequestParam(defaultValue="0")int page,@RequestParam(defaultValue="20")int size){return service.following(id,PageRequests.of(page,size,100));}private UUID user(Jwt j){return UUID.fromString(j.getSubject());}}